Move away from using sub-models.
This was causing some weird graph structures. I'm not sure if this
is an artifact with graph visualization in TensorBoard or a legitimate
issue, but I got rid of the sub-models anyway. It's possible this will
fix the training problem I'm having.

Also, removing sub-models actually makes the code simpler, so I guess
they weren't such a brilliant idea in the first place.